Separated main and forum's html and reworked code to know which to load
