+++ forum.html
@@ -13,7 +13,7 @@
 </head>
 <body>
 
-<div id="MdlApp" title="Main"></div>
+<div id="MdlApp" title="Forum"></div>
 
 <script type="text/javascript" src="out/production/School-Web-Project/lib/kotlin.js"></script>
 <script type="text/javascript" src="out/production/School-Web-Project/School-Web-Project.js"></script>
