--- src/forum.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Foro</title>
-</head>
-<body>
-Foro
-</body>
-</html>
