--- src/site/pages/Forums.kt
@@ -1,20 +1,0 @@
-package site.pages
-
-import components.ListIem
-import components.layout.Grid
-import components.layout.grid
-import components.layout.navigationlayout.Content
-import components.layout.navigationlayout.MdlContent
-import components.layout.navigationlayout.content
-import components.list
-
-
-/**
- * not being used right now.
- * */
-object Forums : MdlContent {
-    override val content = content("Foros") {
-        setAttribute("style", " background: url('images/whoiam.jpg') center / cover; filter: alpha(opacity=60); padding: 10px;")
-    }
-}
-
